requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pillow==10.2.0
pillow-heif==0.15.0
rawpy==0.19.0
imageio==2.34.0
python-multipart==0.0.9
jinja2==3.1.3
aiofiles==23.2.1
